refactor: move default mathFractionDigits and color precision constants in separate constants file
