+++ src/color-modifiers/transformColorModifiers.ts
@@ -3,11 +3,7 @@ import { usesReferences } from 'style-dictionary/utils';
 import { modifyColor } from './modifyColor.js';
 import { ColorModifier } from '@tokens-studio/types';
 import { ColorModifierOptions } from '../TransformOptions.js';
-import { defaultFractionDigits } from '../register.js';
-
-// 3 for backwards compatibility, but should better be 5, referring to the default precision (5) of the colorJS.io library
-// see: https://colorjs.io/docs/output.html#getting-a-string-representation-of-a-color
-export const defaultColorPrecision = 3;
+import { defaultFractionDigits, defaultColorPrecision } from '../utils/constants.js';
 
 /**
  * Helper: Transforms color tokens with tokens studio color modifiers
