+++ src/utils/constants.ts
@@ -1,0 +1,5 @@
+export const defaultFractionDigits = 4;
+
+// 3 for backwards compatibility, but should better be 5, referring to the default precision (5) of the colorJS.io library
+// see: https://colorjs.io/docs/output.html#getting-a-string-representation-of-a-color
+export const defaultColorPrecision = 3;
